checkpoint for logs
