+++ src/command_processor.rs
@@ -163,7 +163,6 @@ impl CommandProcessor {
             Box::new(crate::cmd::ConfigMaps::new()),
             Box::new(crate::cmd::Secrets::new()),
             Box::new(crate::cmd::Jobs::new()),
-            Box::new(crate::cmd::Logs::new()),
             Box::new(crate::cmd::Describe::new()),
             Box::new(crate::command::alias::Alias::new()),
             Box::new(crate::command::alias::Unalias::new()),
@@ -179,6 +178,7 @@ impl CommandProcessor {
             Box::new(crate::command::deployments::Deployments::new()),
             Box::new(crate::command::events::Events::new()),
             Box::new(crate::command::exec::Exec::new()),
+            Box::new(crate::command::logs::Logs::new()),
             Box::new(crate::command::namespaces::Namespace::new()),
             Box::new(crate::command::namespaces::Namespaces::new()),
             Box::new(crate::command::nodes::Nodes::new()),
